requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Caching==2.0.2
Flask-Limiter==3.3.0
Flask-Mail==0.9.1
email-validator==2.0.0
python-dotenv==1.0.0
bleach==6.0.0
Pygments==2.15.1
pytest==7.3.1
gunicorn==20.1.0